+++ CHANGELOG.md
@@ -3,6 +3,7 @@ master
 
 * Can print QR code for blockchain accounts (buchfink list -o qrcode)
 * Can use own ETH RPC for blockchain queries
+* Support Bitcoin Cash
 
 0.0.13
 ------